requirements.txt
boto3==1.9.42
botocore==1.12.42
docutils==0.14
jmespath==0.9.3
python-dateutil==2.7.5
PyYAML==3.13
s3transfer==0.1.13
six==1.11.0
urllib3==1.24.1
